@@ -14,46 +14,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/core": "^6.0.2",
+    "@angular/animations": "~8.2.5",
+    "@angular/cdk": "^8.2.0",
+    "@angular/common": "~8.2.5",
+    "@angular/compiler": "~8.2.5",
+    "@angular/core": "~8.2.5",
+    "@angular/forms": "~8.2.5",
     "@angular/material": "^8.2.0",
+    "@angular/platform-browser": "~8.2.5",
+    "@angular/platform-browser-dynamic": "~8.2.5",
+    "@angular/router": "~8.2.5",
     "@types/gapi": "0.0.36",
     "@types/gapi.auth2": "0.0.47",
     "@types/gapi.client.drive": "^3.0.3",
-    "angular": "^1.7.8",
-    "angular-oauth2-oidc": "^4.0.3",
-    "core-js": "^2.6.9",
+    "angular-oauth2-oidc": "^8.0.4",
+    "core-js": "^3.2.1",
     "d3": "^5.12.0",
     "hammerjs": "^2.0.8",
     "jsmediatags": "^3.9.2",
-    "pouchdb": "^6.4.3",
+    "pouchdb": "^7.1.1",
     "pouchdb-adapter-idb": "^7.1.1",
-    "pouchdb-find": "^6.4.3",
+    "pouchdb-find": "^7.1.1",
     "rxdb": "^8.4.0",
-    "rxjs": "^6.5.3",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
     "wavesurfer.js": "^3.0.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/cli": "^8.3.4",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
+    "@angular-devkit/build-angular": "~0.803.4",
+    "@angular/cli": "~8.3.4",
+    "@angular/compiler-cli": "~8.2.5",
+    "@angular/language-service": "~8.2.5",
     "@types/d3": "^5.7.2",
-    "@types/jasmine": "^2.8.16",
-    "@types/jasminewd2": "^2.0.6",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/jsmediatags": "^3.9.0",
     "@types/node": "~8.9.4",
-    "codelyzer": "^5.1.0",
+    "codelyzer": "^5.0.0",
     "electron": "^6.0.9",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
