@@ -29,12 +29,12 @@ const { shell } = window.require("electron").remote;
 export class TableviewComponent implements OnInit {
   @Input("isDisplayImage")
   isDisplayImage = true;
-  @ViewChild(MatMenuTrigger)
+  @ViewChild(MatMenuTrigger, {static: false})
   trigger: MatMenuTrigger;
-  @ViewChild("table")
+  @ViewChild("table", {static: false})
   table;
 
-  @ViewChild(MatPaginator)
+  @ViewChild(MatPaginator, {static: false})
   paginator: MatPaginator;
 
   // @Input("page")
@@ -324,6 +324,6 @@ export class TableviewComponent implements OnInit {
     return item._id;
   }
 
-  @ViewChild("filter")
+  @ViewChild("filter", {static: false})
   filter: ElementRef;
 }
