@@ -1,7 +1,10 @@
 import { Injectable, EventEmitter } from "@angular/core";
 import { AppRepository } from "../repositories/app.repository";
 import { ignoreElements } from "rxjs/operators";
-
+const CLIENT_ID = "584574268984-ia9hs8a653a8p4q23duqob501a5f3fgq.apps.googleusercontent.com";
+ const API_KEY = "AIzaSyADt_kqAFO0hcQ1wKY2jw20LMjDVKc8flo";
+ const DISCOVERY_DOCS = ["https://www.googleapis.com/discovery/v1/apis/drive/v3/rest"];
+ const SCOPES = 'https://www.googleapis.com/auth/drive';
 
 @Injectable()
 export class GapiSession {
